requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
pyodbc>=4.0.39

# Configuration and environment
python-dotenv>=1.0.0
PyYAML>=6.0

# Logging and utilities
colorlog>=6.7.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Excel/reporting (for future use)
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Fuzzy matching
rapidfuzz>=3.6.1

# HTTP requests (for API integrations)
requests>=2.31.0

# Data validation
cerberus>=1.3.4

# Data visualization (optional, for future use)
matplotlib>=3.7.0

# Status reports
tqdm>=4.66.1
